requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
pillow>=9.5.0
requests>=2.28.0
torch>=2.0.0
torchvision>=0.15.0

